商城管理服务端实现

+++ gms-server/src/main/java/org/gms/dao/entity/ModifiedCashItemDO.java
@@ -41,7 +41,7 @@ public class ModifiedCashItemDO implements Serializable {
     /**
      * 数量
      */
-    private Integer count;
+    private Short count;
 
     /**
      * 价格
